requirements.txt
kucoin-futures-python == 1.0.6
python-dotenv == 0.19.2
pytz == 2021.3